+++ app/src/main/kotlin/com/softteco/template/utils/bluetooth/BluetoothByteParserImpl.kt
@@ -4,6 +4,7 @@ import com.softteco.template.Constants
 import com.softteco.template.data.bluetooth.BluetoothByteParser
 import com.softteco.template.data.bluetooth.entity.BluetoothDeviceData
 import com.softteco.template.data.bluetooth.entity.BluetoothDeviceType
+import com.softteco.template.utils.characteristicByteConversation
 import javax.inject.Inject
 import javax.inject.Singleton
